refactor: Code update to .NET 6 template format.
